# docs/swagger.yaml
basePath: /v1
definitions:
  candidate.AddCertificationRequest:
    properties:
      certification_name:
        type: string
      expiration_date:
        type: string
      issue_date:
        type: string
      issued_by:
        type: string
    required:
    - certification_name
    - issue_date
    - issued_by
    type: object
  candidate.AddEducationRequest:
    properties:
      degree:
        type: string
      description:
        type: string
      end_date:
        type: string
      institution:
        type: string
      start_date:
        type: string
    required:
    - degree
    - institution
    - start_date
    type: object
  candidate.AddExperienceRequest:
    properties:
      company:
        type: string
      description:
        type: string
      end_date:
        type: string
      job_title:
        type: string
      start_date:
        type: string
    required:
    - company
    - job_title
    - start_date
    type: object
...
  response.Response:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
      status:
        type: string
    type: object
host: dz-jobs-api-production.up.railway.app
info:
  contact: {}
  description: This is the API documentation for the DZ Jobs portal.
  title: DZ Jobs API
  version: "1.0"
paths:
...
  /candidates/{candidate_id}/certifications:
    get:
      description: Get all certifications for a candidate by candidate ID
      parameters:
      - description: Candidate ID
        in: path
        name: candidate_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
      summary: Get certifications by candidate ID
      tags:
      - Candidates - Certifications
  /candidates/{candidate_id}/certifications/{certification_id}:
    delete:
      description: Delete a certification by candidate ID and certification ID
      parameters:
      - description: Candidate ID
        in: path
        name: candidate_id
        required: true
        type: string
      - description: Certification ID
        in: path
        name: certification_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
      summary: Delete certification
      tags:
      - Candidates - Certifications
  /candidates/{candidate_id}/education:
    delete:
      description: Delete an education record by candidate ID
      parameters:
      - description: Candidate ID
        in: path
        name: candidate_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
      summary: Delete education record
      tags:
      - Candidates - Education
    get:
      description: Get all education records for a candidate by candidate ID
      parameters:
      - description: Candidate ID
        in: path
        name: candidate_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
      summary: Get education records by candidate ID
      tags:
      - Candidates - Education
    post:
      consumes:
      - application/json
      description: Add a new education record for a candidate
      parameters:
      - description: Candidate ID
        in: path
        name: candidate_id
        required: true
        type: string
      - description: Education request
        in: body
        name: education
        required: true
        schema:
          $ref: '#/definitions/candidate.AddEducationRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
      summary: Add a new education record
      tags:
      - Candidates - Education
  /candidates/{candidate_id}/experience:
    delete:
      description: Delete an experience record by candidate ID
      parameters:
      - description: Candidate ID
        in: path
        name: candidate_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
      summary: Delete experience record
      tags:
      - Candidates - Experience
    get:
      description: Get all experience records for a candidate by candidate ID
      parameters:
      - description: Candidate ID
        in: path
        name: candidate_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
      summary: Get experience records by candidate ID
      tags:
      - Candidates - Experience
    post:
      consumes:
      - application/json
      description: Add a new experience record for a candidate
      parameters:
      - description: Candidate ID
        in: path
        name: candidate_id
        required: true
        type: string
      - description: Experience request
        in: body
        name: experience
        required: true
        schema:
          $ref: '#/definitions/candidate.AddExperienceRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
      summary: Add a new experience record
      tags:
      - Candidates - Experience
...
  /candidates/{id}/certifications:
    post:
      consumes:
      - application/json
      description: Add a new certification for a candidate
      parameters:
      - description: Candidate ID
        in: path
        name: id
        required: true
        type: string
      - description: Certification request
        in: body
        name: certification
        required: true
        schema:
          $ref: '#/definitions/candidate.AddCertificationRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
      summary: Add a new certification
      tags:
      - Candidates - Certifications
...
swagger: "2.0"